+++ packages/core/src/components/highlight-box/highlight-box.template.ts
@@ -1,6 +1,7 @@
 import { HighlightBox } from '@dso-toolkit/sources';
 import { html, TemplateResult } from 'lit-html';
 import { ifDefined } from 'lit-html/directives/if-defined';
+import { unsafeHTML } from 'lit-html/directives/unsafe-html';
 
 import { iconTemplate } from '../icon/icon.template';
 
@@ -14,7 +15,9 @@ export function highlightBoxTemplate({ yellow, white, dropShadow, border, step, 
       ?border=${border}
     >
       ${icon && iconTemplate({ icon }, 'icon')}
-      ${richContent}
+      <div class="dso-rich-content">
+        ${typeof richContent === 'string' ? unsafeHTML(richContent) : richContent}
+      </div>
     </dso-highlight-box>
   `;
 }