+++ packages/css/src/components/row-equal-heights/row-equal-heights.template.ts
@@ -1,0 +1,10 @@
+import { RowEqualHeights } from '@dso-toolkit/sources';
+import { html, TemplateResult } from 'lit-html';
+
+export function rowEqualHeightsTemplate({ children }: RowEqualHeights<TemplateResult>) {
+  return html`
+    <div class="row dso-equal-heights">
+      ${children}
+    </div>
+  `;
+}