+++ packages/sources/src/components/row-equal-heights/row-equal-heights.models.ts
@@ -1,0 +1,3 @@
+export interface RowEqualHeights<TemplateFnReturnType> {
+  children: TemplateFnReturnType;
+}